Change thread pool used for TTL-based GC in ZkBucketDataAccessor
ZkBucketDataAccessor supports TTL-based garbage collection of stale versions. We create a thread pool for this of size 1, but the way the thread pool is created currently may not guarantee strictly sequential execution of GC timer tasks, which may cause unexpected GC behaviors. This commit uses Executors.newSingleThreadScheduledExecutor()to create the GC thread pool with a sequential execution guarantee.