Change thread pool used for TTL-based GC in ZkBucketDataAccessor (#945)
ZkBucketDataAccessor supports TTL-based garbage collection of stale versions. We create a thread pool for this of size 1, but the way the thread pool is created currently may not guarantee strictly sequential execution of GC timer tasks, which may cause unexpected GC behaviors. This commit uses Executors.newSingleThreadScheduledExecutor()to create the GC thread pool with a sequential execution guarantee.

Note: another way to do this is to implement a Bounded event queue-based GC, but Java already provides this queue-based model in Executors interface - we want to try our best at avoid reinventing the wheel.