Fix test testDropInstance (#1053)
Test testDropInstance doesn't assert expected HelixException, so when a live instance is not created, the test still passes. The reason is the test uses a sharedZkClient to create a live instance but fails.

This PR addresses this issue by creating a dedicated zkclient in ZkBaseDataAccessor to create a live instance.