Fix the issue that the instance may not be assigned a replica as expected.
This is to fix a regression which was introduced by PR #986.
The PR tried to prioritize the preference list to avoid unnecessary top state transition. However, there was a bug in the prioritizing logic and if one participant is skipped due to low priority, it won't be picked up again during the cauculating. As a result, this participant won't be assigned with any replica even it is originally in the preference list.
This fix will ensure the skipped participant being checked again until it gets the assignment.