Use dedicated ZkClient in HelixPropertyFactory
It's been observed that sometimes using a shared Zk connection could cause issues when we write in-memory tests that shut down ZK server and restart it. This is because with a shared Zkconnection, we may see the client's zxid be higher than that of the server, which is an invalid state in ZooKeeper. Since HelixManager is a single-realm dedicated concept, it is okay to use a dedicated zk connection to pull CloudConfig in getHelixManagerProperty() method. This will solve issues caused by zxid mismatches.