Shutdown the TaskStateModelFactory threads created in the tests.
The TaskStateModelFactory initialization creates threadpool. The expectation is that the application code closes the threadpool and the threads when the participant instance is shuting down. In most cases, this means the JVM is going to be shutdown. So this operation is not a must. However, in the test cases, these threadpools leak thousands of the threads.
This PR adds cleanup logic to shutdown threadpools that are created for the partcipant instances. Note that there are still threads leakage when the participants are created separately instead of using the general methods.