Fix validation logic in GenericApiBuilders and add usage tests (#1127)
Some of the validation logic in GenericBaseDataAccessorBuilder and GenericZkHelixApiBuilder were confusing and there were no usage tests. This commit reviews the consistency between validate methods of BaseDataAccessor APIs and writes test cases to demonstrate how they should be used.

Also, connection timeout setting wasn't being honored in the createZkClient() methods, so this bug was fixed as well.

createZkConnectionConfig() was removed from RealmAwareZkClient because the intended purpose of this method isn't clear and was not being used anywhere else.