Fix TestRebalancePipeline test (#1157)
In the test testDuplicateMsg from TestRebalancePipeline, currently an invalid session id is used.
The wrong session id makes the following test and assertion meaningless. This PR uses the correct session id liveInstances.get(0).getEphemeralOwner() instead.

Also, we found a NPE thrown during the pipeline, causing the pipeline to stop, due to the missing thread pool in setAsyncTasksThreadPool. This PR works around the issue by giving it a thread pool and adds an issue reference in the comment block.