ZkClient should not keep retrying getChildren() due to large number of children (#1109)
For ZkClient's getChildren() operation, if there are a large number of children and the response packet size exceeds jute.maxbuffer default value 4MB on zk client side, ZkClient will get a ConnectionLossException and keep retrying connecting to ZK. The consequence is, the infinite retry may cause heavy GC on ZK server and kill ZK server.

This commit implements a workaround to exit retry loop for getChildren() if a large number of children cause connection loss.