Implement thread leakage check for each test class #1226
Recently unit test can not complete due to threads leaking
and ZK session leaking. We identified the leaks and fixed
them. In order for future test not to leak resources, we
need a way to enforce test class would release/clean up resource created during test.

A check is implemented in @afterclass in ZkTestBase.java.