Validate instance topology configuration before let it comes online  (#1129)
Helix does not check if an instance's topology config is defined correctly when the instance is added or joins cluster. This will cause Helix to fail the rebalance. This change adds the sanity check when

Instance joins cluster as liveInstance
User update InstanceConfig using RestApi
Controller prepares the topology structure for rebalance
This change also did come code clean up and centralized the logic for topology sanity check.