Extend the task future tracking to a Map instead of a single object.
Since the ZkBucketDataAccessor is used to access multiple paths, a single task future track is not enough. This change uses a map with the path as the key to track tasks for different paths.
Also enhance the test case to cover the scenario of multiple paths being accessed in a short period.