Add HelixManager constructor with RealmAwareZkConnectionConfig (#1183)
Previously, there wasn't a way to use HelixManager with a custom routing data ZK connection config. This new way of constructing HelixManager allows users to use HelixManager with a custom RealmAwareZkConnectionConfig.

This PR also adds a null check for zkAddress for existing Helix APIs so that it will know when to start themselves in a multi-zk mode.