Use the DelayedAutoRebalancer as the default rebalancer.
If no rebalancer class name is specified, use the DelayedAutoRebalancer instead of AutoRebalancer.

Note that DelayedAutoRebalancer is compatible with the AutoRebalanceStrategy. So the assignment would remain the same even after migration.

However, since DelayedAutoRebalancer keeps a cached rebalance result in the Controller cache, it causes a behavior change.
- We need to modify multiple test classes to use the CrushEd rebalance strategy for the correct test logic.
- This behavior change only impacts BestPossibleExternalViewVerifier if the resource was on the older AutoRebalancer (by DEFAULT setting) and AutoRebalanceStrategy.
- If any Helix client is using BestPossibleExternalViewVerifier with the described configuration, then changing the rebalancer class config in the IdealState explicitly to AutoRebalancer is required before the Helix version bump-up. Or the BestPossibleExternalViewVerifier may fail.

If the client application code is using BestPossibleExternalViewVerifier with AutoRebalancer/AutoRebalanceStrategy, then to avoid problem about this change, please do either of the following 2 changes.
- Explicitly configure the AutoRebalancer class in the resource IdealState.
- Change to use StrictMatchExternalViewVerifier. Note that using this verifier requires PERSIST_BEST_POSSIBLE_ASSIGNMENT to be true in the cluster config.