Enhance ZkHelixVerifier and its related subclass to take a default wait period (#1328)
We have random sleep() in the test before verifyByPolling().
This causes test failure in different environment. Here we enhance ZKHelixVerifier
family with an option in builder to add a waiting period. To keep it backward
compatible, without enable this waiting option, the code would behave exactly the
same as before.