Add metric annotations to each resource method (#1379)
To enable availability monitoring for helix rest, we want request latency and response metrics. And make sure metrics are not consuming too much memory, we want to categorize the resource methods into "read" and "write". GET verb is read, and POST, DELETE, UPDATE are writes.
This commit only adds @Timed and @ResponseMetered annotations to each rest method and categorize them "read" or "write".