Stabilize the TestJobTimeout (#1398)
In this commit, TestJobTimeout has been stabilized by increasing
the job's timeout. The tasks will get ABORTED if the job is TIMING_OUT
and the tasks have already been started. Since 10ms is very short, it is
possible that the job goes to the TIMED_OUT state without tasks being even
started on the participants. Hence, a longer timeout time will stabilize
the test.