Add TrieClusterTopology for retrieving hierarchical topology (#1307)
Add TrieNode class to define a node in the trie
Add ClusterTrie class to handle the construction, validation and retrieval of nodes/paths in the trie.
Add ClusterTopology class to provide different APIs for users to retrieve cluster topology information.
Add APIs in HelixAdmin to retrieve ClusterTopology of a specific cluster.