HelixClusterVerifier verify() with default waitTillVerify time -- part one (#1449)
HelixClusterVerifier verify() and related method may return
prematurely. The reason is that verify the converging stable
condition too early before the controller has a chance to make
a change. Basically, the previous stable state is mistaken as the
expected next stable state.
This commit fixes this issue by adding waitTillVerify() timeout in
the construction time of the verifier.