Double check if the HelixTaskExecutor is shutting down before make Partition in ERROR state when schedule task fails. (#1515)
There is a race condition between TaskExecutor thread pool shutting down and Message handler stops listening.
In this gap, the message will still be processed but the schedule will fail.
If we mark partition into an ERROR state, then the controller side logic might be confused.
This PR adds an additional check to avoid the unnecessary ERROR state. But please note this is a workaround instead complete fix.