Pass partitions list to custom client rest (#1570)
Problem:
CustomRestClientImpl serializes the partitions list to a string before sending the payloads to the custom client health check Rest API. The custom client is expecting a list of partitions, but received a string.
So the custom client returns an error 400. But actually the partitions are healthy. This is caused by the serialization/deserialization inconsistency between helix rest and custom client rest.

Solution:
Keep the partitions list in the payload without serializing it to a string when sending post request. So the custom client will get the list of partitions after deserializing the payloads.