Update the internal fields of the Data Providr when overriding the data through set methods.
It was observed that the set methods, which are used but the Helix tools or tests, are not updating the derived fields.
This makes the override incomplete and it causes some potential issues when the tools are used.
This PR ensures that the set methods are modifying important derived fields.