Update pull_request_template.md
As discussed, we are now relying on the Github CI test. So the local mvn test is no longer a requirement.
I didn't remove the section since it is possible that the CI test may not run due to some unexpected reasons. In such a case, the guideline is to specify the known issue and keep running the test locally.