Fix TestCrushAutoRebalanceNonRack.testLackEnoughInstances unstable issue (#1630) (#1631)
This test still fails in production due to the race condition that the controller sending message is async
to test dropping instance. Thus when there is a message on the queue, dropping would throw
an exception. Here the test is stabilized by stopping the controller as a workaround to avoid such race condition.