Directly use Best Possible State to calculate DifferenceWithIdealStateGauge metrics instead of relying on the persisted IdealState. (#1697)
This PR aims to resolve the undesired dependencies between DifferenceWithIdealStateGauge metric and the PERSIST_XXXXXX_ASSIGNMENT configuration. Before the change, if the assignment is not persisted, then the metric will report incorrect data.
To fix the issue, the calculation must be moved from the ExternalViewComputeStage to the BestPossibleStateCalcStage. Because the Best Possible State is only available in the BestPossibleStateCalcStage assuming the PERSIST_XXXXXX_ASSIGNMENT option is not turned on. In addition to the main logic migration, other changes listed following are required to ensure multi-thread safty.
1. Concurrent control the EV cache since it will now be read in the BestPossibleStateCalcStage in addition to the ExternalViewComputeStage.
2. Minor changes in the diff computing logic for the corner cases since the Best Possible State state mapping is not exactly the same as IdealState persist assignment.
3. Cleanup PersistAssignmentStage logic so it won't modify the IdealState cache anymore. It is supposed to be read-only.
4. Test cases are modified to cover the changes. Also the metric is now reporting correct result in some corner cases such as all the nodes are disabled. The test cases which are testing based on the wrong behavior have been fixed.