Fix unexpected result when resuming a cluster from paused/maintenance mode.
This PR aims to fix a race condition that the resume event could be pushed to the event queue before the cluster status has been updated. In this case, when the event is processed, the controller might still think the cluster is in paused/maintenance mode. As a result, the resume event is discard and the rebalance won't be done until the next event happens.