Change MessageQueueMonitor from static to dynamic metric (#1715)
* Change MessageQueueMonitor from static to dynamic metric

This commit modifies the MessageQueueMonitor class to use dynamic
metrics by extending the DynamicMBeanProvider class. It also
deprecates the MessageQueueMonitorMBean interface because it is
no longer used..

After converting MessageQueueMonitor from static to dynamic metric, the
MessageQueueMonitorMBean interface is no longer used in the code. Since the
goal is to remove all direct references to the MBean class, this interface
is deleted.