Remove the logic that disables the partitions which fails to be state-transited from the ERROR state. (#1721)
This logic of disabling an ERROR partition with additional failure transition causes confusion. When the partitions are disabled unexpected, the system admin has to double-check and manually enable the partitions.
In reality, if the partition fails to finish a state transition from the ERROR state, it will remain in the ERROR state. And disabling it in addition is not desired. Otherwise, if it is eventually reset successfully, it should be good to continue serving. Sanity check shall be done in the ERROR->OFFLINE state transition process.

[Behavior Changed] This PR changes the existing behavior.