Change instance status in Helix UI to be colorblind-friendly (#1759)
This commit changes the instance online/offline status colors in Helix UI from green/red to blue(#2196F3)/black(#212121) to be more colorblind-friendly. The colors are chosen such that they have a contrast ratio of at least 3:1 against each other and the background. Moreover, a checkmark is placed in the icon for online status, and a cross (x) for the offline.