Change partition status in Helix UI to be colorblind-friendly (#1785)
This commit changes the partition status colors in Helix UI from
green/orange to blue(#2196F3)/black(#212121) to be more colorblind-
friendly. The colors are chosen such that they have a contrast ratio
of at least 3:1 against each other and the background. Moreover, a
checkmark is placed in the icon for ready status, and an exclamtion
point for the not-ready status.