Fix flaky test testClusterFreezeMode (#1871)
The test fails randomly. Root cause is the test used a random cluster from the clusters set which has super clusters and task cluster, so the cluster could be a task cluster and then the test fails. Freeze mode does not apply to task framework.

This commit fixes it by changing the test cluster name with a fixed name TestClusters_0.