Fix failing ZkClient TTL test (#2120)
Some tests for the TTL feature rely on zookeeper.extendedTypesEnabled system property to be set. However, interleaving of test methods from different classes resulted in the property being cleaned by one completed test class while another class has tests that rely on this property still running. This commit sets the property at the beginning of every method that needs it, and cleans at the end of the method.