Fix flaky integration test for helix-rest for TestClusterAccessor#testClusterFreeze. (#2451)
Fixing Flaky Integration Test TestClusterAccessor#testClusterFreeze

This integration test is making a POST request to helix-rest to place cluster TestCluster_0 into CLUSTER_FREEZE status. It then attempts to verify that the following GET request for clusterStatus is in CLUSTER_FREEZE mode.

Previously we were running this verification to check if the clusterStatus znode existed. This will not necessarily indicate that the pause event has been processed, as there could be an earlier event passed through the _managementEventPipeline that causes the znode to be created with clusterMode == NORMAL. This would lead us to make the GET request too early.

Instead, we will verify that the clusterStatus znode has CLUSTER_FREEZE mode before we make the GET request. This will ensure that the pause event was already processed.