WAGED rebalance overwrite redesign -- part 1 (#2444)
Implement new code path to handle "rebalance overwrites" that honors WAGED hard constraints
The rebalance overwrite is a special mechanism to handle minActiveReplica during delayed rebalance. Additional replicas might be brought up in an instance that violates the WAGED hard constraint. This commit implements new code logic for it that honors the hard constraints. The new code is NOT yet integrated to existing flow.