Fix for PR-2481 - Disconnection method can throw exception.
The testcase issues 'disconnect()' from a thread and in the main-thread
waits for the status to get updated. It is possible that in disconnect() code
path, there can be an exception and we never get to setting the status.
Next Assert() is in place to see if we actually disconnected or not.

So at least pass the first Assert() as it is created by test logic and
not actual code logic.

Testing Done:
mvn test logs
--------------
[ERROR] Failures:
[ERROR]   TestNoThrottleDisabledPartitions.testDisablingTopStateReplicaByDisablingInstance:98 expected:<false> but was:<true>
[ERROR]   TestP2PNoDuplicatedMessage.testP2PStateTransitionEnabled:180 expected:<true> but was:<false>
[ERROR]   TestParticipantFreeze.testUnfreezeParticipant:228 expected:<true> but was:<false>
[ERROR]   TestWagedRebalanceTopologyAware.testAddZone:112->TestWagedRebalanceFaultZone.testAddZone:270->TestWagedRebalanceFaultZone.validate:318->TestWagedRebalanceFaultZone.validateZoneAndTagIsolation:342 expected:<3> but was:<2>
[ERROR]   TestRecurringJobQueue.testDeletingRecurrentQueueWithHistory:298 expected:<true> but was:<false>
[ERROR]   TestResourceThreadpoolSize.testBatchMessageThreadPoolSize:206 expected:<true> but was:<false>
[ERROR] Tests run: 1326, Failures: 6, Errors: 0, Skipped: 0