MissingTopState when the leader comes back on the same host
We currently don't handle the case where if the user has only one replica and the leader partition comes back on the same host as there are no other replica.

Testing done:
Added a test scenario in the JSON config file where the partition is brought oneline on the same host and validated that the test passes.