Fix CI: TestWorkflowTermination::testWorkflowPausedTimeout
This test case failed just once and that too recently. We haven't changed
any core logic in the code path of Task management.

The failure happened because:
2023-06-05T21:37:59.7170747Z [ERROR] testWorkflowPausedTimeout(org.apache.helix.integration.task.TestWorkflowTermination)  Time elapsed: 10.13 s  <<< FAILURE!
2023-06-05T21:37:59.7171834Z org.apache.helix.HelixException: Fail to stop the workflow/queue testWorkflowPausedTimeout with in 10000 milliseconds.

Indicating the timeout was not sufficient. So increased the timeout.

Testing done: since this is CI failure, ran into my private branch 3/4 times
https://github.com/desaikomal/helix/actions/workflows/Helix-PR-CI.yml